+++ tests/ipython/test_ipython.py
@@ -2,10 +2,8 @@ from pathlib import Path
 
 import pytest
 from IPython.core.error import UsageError
-from IPython.testing.globalipapp import get_ipython
 
 from kedro.framework.project import pipelines
-from kedro.framework.startup import ProjectMetadata
 from kedro.ipython import (
     _find_node,
     _load_node,
@@ -17,73 +15,9 @@ from kedro.ipython import (
     magic_load_node,
     reload_kedro,
 )
-from kedro.pipeline import node
 from kedro.pipeline.modular_pipeline import pipeline as modular_pipeline
 
-PACKAGE_NAME = "fake_package_name"
-PROJECT_NAME = "fake_project_name"
-PROJECT_INIT_VERSION = "0.1"
-
-
-@pytest.fixture(autouse=True)
-def cleanup_pipeline():
-    yield
-    from kedro.framework.project import pipelines
-
-    pipelines.configure()
-
-
-@pytest.fixture(scope="module", autouse=True)  # get_ipython() twice will result in None
-def ipython():
-    ipython = get_ipython()
-    load_ipython_extension(ipython)
-    return ipython
-
-
-@pytest.fixture(autouse=True)
-def fake_metadata(tmp_path):
-    metadata = ProjectMetadata(
-        source_dir=tmp_path / "src",  # default
-        config_file=tmp_path / "pyproject.toml",
-        package_name=PACKAGE_NAME,
-        project_name=PROJECT_NAME,
-        kedro_init_version=PROJECT_INIT_VERSION,
-        project_path=tmp_path,
-        tools=None,
-        example_pipeline=None,
-    )
-    return metadata
-
-
-@pytest.fixture(autouse=True)
-def mock_kedro_project(mocker, fake_metadata):
-    mocker.patch("kedro.ipython.bootstrap_project", return_value=fake_metadata)
-    mocker.patch("kedro.ipython.configure_project")
-    mocker.patch("kedro.ipython.KedroSession.create")
-
-
-def dummy_function(dummy_input, my_input):
-    """
-    Returns True if input is not
-    """
-    # this is an in-line comment in the body of the function
-    random_assignment = "Added for a longer function"
-    random_assignment += "make sure to modify variable"
-    return not dummy_input
-
-
-@pytest.fixture
-def dummy_function_defintion():
-    body = '''def dummy_function(dummy_input, my_input):
-    """
-    Returns True if input is not
-    """
-    # this is an in-line comment in the body of the function
-    random_assignment = "Added for a longer function"
-    random_assignment += "make sure to modify variable"
-    return not dummy_input
-'''
-    return body
+from .conftest import dummy_function
 
 
 def dummy_nested_function(dummy_input):
@@ -97,28 +31,6 @@ def dummy_function_with_loop(dummy_list):
     for x in dummy_list:
         continue
     return len(dummy_list)
-
-
-@pytest.fixture
-def dummy_node():
-    return node(
-        func=dummy_function,
-        inputs=["dummy_input", "extra_input"],
-        outputs=["dummy_output"],
-        name="dummy_node",
-    )
-
-
-@pytest.fixture
-def lambda_node():
-    return node(
-        func=lambda x: x,
-        inputs=[
-            "x",
-        ],
-        outputs=["lambda_output"],
-        name="lambda_node",
-    )
 
 
 @pytest.fixture
